+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 9.6.1 - *2021-09-20*
+- Move the docker_image library to a custom resource
 
 ## 9.6.1 - *2021-09-20*